21-5-2021

+++ lib/Model/ProductRateModel.dart
@@ -46,7 +46,7 @@ class ProductRateDetail {
 
   String id;
   String comment;
-  int rateNum;
+  double rateNum;
   String productId;
   String photo;
   String userId;
